# WebGPURenderer: Compute `modelViewMatrix` using GPU