--- a/examples/webgpu_tsl_procedural_terrain.html
+++ b/examples/webgpu_tsl_procedural_terrain.html
@@ -27,7 +27,7 @@
 		<script type="module">
 
 			import * as THREE from 'three';
-			import { mx_noise_float, color, cross, dot, float, modelNormalMatrix, positionLocal, sign, step, Fn, uniform, varying, vec2, vec3, Loop } from 'three/tsl';
+			import { mx_noise_float, color, cross, dot, float, modelViewMatrix, positionLocal, sign, step, Fn, uniform, varying, vec2, vec3, Loop } from 'three/tsl';
 
 			import { GUI } from 'three/addons/libs/lil-gui.module.min.js';
 			import { OrbitControls } from 'three/addons/controls/OrbitControls.js';
@@ -150,7 +150,7 @@
 			
 				} )();
 
-				material.normalNode = modelNormalMatrix.mul( vNormal );
+				material.normalNode = modelViewMatrix.transformDirection( vNormal );
 
 				material.colorNode = Fn( () => {