--- a/src/nodes/accessors/InstanceNode.js
+++ b/src/nodes/accessors/InstanceNode.js
@@ -1,9 +1,9 @@
 import Node from '../core/Node.js';
 import { varyingProperty } from '../core/PropertyNode.js';
 import { instancedBufferAttribute, instancedDynamicBufferAttribute } from './BufferAttributeNode.js';
-import { normalLocal } from './Normal.js';
+import { normalLocal, transformNormalToView } from './Normal.js';
 import { positionLocal } from './Position.js';
-import { nodeProxy, vec3, mat3, mat4 } from '../tsl/TSLBase.js';
+import { nodeProxy, vec3, mat4 } from '../tsl/TSLBase.js';
 import { NodeUpdateType } from '../core/constants.js';
 import { buffer } from '../accessors/BufferNode.js';
 import { instanceIndex } from '../core/IndexNode.js';
@@ -103,11 +103,7 @@ class InstanceNode extends Node {
 
 		if ( builder.hasGeometryAttribute( 'normal' ) ) {
 
-			const m = mat3( instanceMatrixNode );
-
-			const transformedNormal = normalLocal.div( vec3( m[ 0 ].dot( m[ 0 ] ), m[ 1 ].dot( m[ 1 ] ), m[ 2 ].dot( m[ 2 ] ) ) );
-
-			const instanceNormal = m.mul( transformedNormal ).xyz;
+			const instanceNormal = transformNormalToView( normalLocal, instanceMatrixNode );
 
 			// ASSIGNS