--- a/src/renderers/common/Renderer.js
+++ b/src/renderers/common/Renderer.js
@@ -16,6 +16,8 @@ import QuadMesh from './QuadMesh.js';
 import RenderBundles from './RenderBundles.js';
 import NodeLibrary from './nodes/NodeLibrary.js';
 
+import { modelViewMatrix } from '../../nodes/TSL.js';
+
 import NodeMaterial from '../../materials/nodes/NodeMaterial.js';
 
 import { Scene } from '../../scenes/Scene.js';
@@ -80,7 +82,8 @@ class Renderer {
 		this.info = new Info();
 
 		this.nodes = {
-			library: new NodeLibrary()
+			library: new NodeLibrary(),
+			modelViewMatrix
 		};
 
 		// internals