… examples/screenshots/webgpu_tsl_raging_sea.jpg
… examples/webgpu_tsl_raging_sea.html