Related issue: https://github.com/mrdoob/three.js/pull/28719
Related: https://lxjk.github.io/2017/10/01/Stop-Using-Normal-Matrix.html

**Description**

This change is part of the integration process of https://github.com/mrdoob/three.js/pull/28719 for less CPU usage. These changes brought a gain of around ~25% in performance for scenes with many objects.

| no bundle - old | no bundle - now |
| ------------- | ------------- | 
| 12.10ms | 8.85ms |
| ![image](https://github.com/user-attachments/assets/c13d5dfe-f887-4d8a-a056-ccfdf6cbcf3d) | ![image](https://github.com/user-attachments/assets/4f7f9758-dd5e-4ebe-92bd-2bfa107e2a88) |  

| bundle - old | bundle - now |
| ------------- | ------------- |
| 5.67ms | 3.05ms |
| ![image](https://github.com/user-attachments/assets/4e39ee58-f8d0-4d07-b3fc-70402ada23bf) | ![image](https://github.com/user-attachments/assets/a0143717-2717-478d-82f7-a10b2cf926f8) |