@WestLangley 2024-09-02
We do something similar [here](https://github.com/mrdoob/three.js/blob/598db0df6aa349511eef16b67609012060fc1037/src/renderers/shaders/ShaderChunk/defaultnormal_vertex.glsl.js#L29-L34) for instancing, but this only works if the columns of the matrix are orthogonal, which is not true in general.

See [this explanation](https://github.com/mrdoob/three.js/issues/18497#issuecomment-583771048) -- especially the last sentence.