@sunag 2024-09-02
Thanks @WestLangley! 

Now I just need to check another approach to `webgpu_postprocessing_motion_blur` example :)