@gkjohnson 2024-09-03
I'm not as familiar with shader nodes so I may be misunderstanding what's happening here but I'll write my two cents from what I understand from the description:

This change could cause precision issues when using large coordinates since GPU calculations use 32 bit math which hasn't been an uncommon issue with instances and skinned meshes with large position values (bone and instance matrices are multiplied into the mv matrix on the gpu). See [here](https://discourse.threejs.org/t/the-projectionmatrix-seems-to-have-caused-a-loss-of-accuracy/58455) and [here](https://discourse.threejs.org/t/large-coordinates/50621/7). I suspect it will be even more common if this is done on the GPU for every mesh.

Assume the camera is far from the origin (camera orbiting a to-scale globe model with a radius of 6.3e6 meters) meaning the objects in frame have extremely large positional values, as well. If the MV matrix is calculated on the CPU then 64-bit precision is used meaning any error resulting from the calculations will be much smaller than it will be if these calculations are done with 32 bits on the GPU. This can cause _very_ noticeable jitter artifacts during rendering.