@RenaudRohlinger 2024-09-03
In that case, similar to how `logarithmicDepthBuffer` improves depth precision at the cost of performance, we could introduce another option like `renderer.normalMatrix` (or somethingNormalCPU...) for 64-bit precision when jitter artifacts occur.
This would still allow for high-precision MV matrix calculations on the CPU, would align with the existing `logarithmicDepthBuffer` logic, and improve performances while giving developers an option to increase precision in demanding scenarios.