@sunag 2024-09-03
I like the idea of ​​having a point of origin relative to the camera as presented in item `3.2.1` of this article https://www.diva-portal.org/smash/get/diva2:275843/FULLTEXT02.