@gkjohnson 2024-09-03
> we could introduce another option like `renderer.normalMatrix`

To be clear is this just for normal matrices? This PR involves moving both the model-view matrix multiplication (and implicitly the normal matrix generation) to the GPU, right? In this case we'd want to name it something indicating it's for more than just normal matrices.

> I like the idea of ​​having a point of origin relative to the camera as presented in item 3.2.1 of this article https://www.diva-portal.org/smash/get/diva2:275843/FULLTEXT02.

This is what multiplying the model and view matrices on the CPU is achieving - ie what WebGLRenderer is already doing.