@sunag 2024-09-03
I think the idea is to have global matrices relative to the camera position.