@WestLangley 2024-09-03
Restatement of my previous comment:

The technique proposed in this PR will only be correct when the columns of the model view matrix are orthogonal. The columns will typically not be orthogonal when, for example,

(a) a non-uniformly-scaled parent has a rotated child,

(b) a user-provided object matrix has non-orthogonal columns.