@WestLangley 2024-09-03
Maybe revisit https://github.com/mrdoob/three.js/issues/5974, instead.