@aardgoose 2024-09-03
I have been experimenting with a similar ideas (obviously restricted to uniform scaling) but made an opt in to allow  object.static as proposed in #28719. Thus the existing known to be correct behavior is preserved but a lighter CPU varient is available for renderBundles - (to get lighting working, light uniforms need moving into a shared bindGroup etc).

https://github.com/aardgoose/three.js/tree/freeze2