@sunag 2024-09-03
This would not use matrix multiplication, it would be a simple subtraction of the objects' world matrix position with the camera's world position, in which case the camera world would always have zero position for the GPU. It is certainly something else to add for CPU how to calculate `matrixView` and `normalMatrix` is today, not ideal for Bundler, but since `three.js` currently does not have dedicated API for "huge open world", given the issues you presented in `WebGLRenderer`, the `viewMatrix` calculated on the GPU does not solve problems such as Attached `SkinnedMesh`, `InstacedMesh` and probably others that a world matrices relative to the camera position should resolve.

It is also possible to notice that most of the issues are related to the incorrect use of the scale, where `1 meter` should be `1.0`.

I don't think there is a perfect solution in this here, just since this PR is prioritizing performance and keeping it functional in situations where the camera needs to move `10 kilometers` of distance of center of scene, which seems reasonable to me. 

We could have Nodes to deal with these situations since the function call is deferred we would not have problems in defining how the `viewMatrix` is constructed for a given object, these are other possibilities to be studied.