@sunag 2024-09-03
> We could have Nodes to deal with these situations since the function TSL Fn call are deferred we would not have problems in defining how the viewMatrix is constructed for a given object, these are other possibilities to be studied.

It seems like the best way to close this issue:
You can use `highPrecisionModelViewMatrix` for global or specific cases for example:


Global usage:
```js
// global
import { highPrecisionModelViewMatrix } from 'three/tsl';

const renderer = new THREE.WebGPURenderer( { antialias: true } );
renderer.nodes.modelViewMatrix = highPrecisionModelViewMatrix; // it will replace all MVP with this modelView
```

Single Material / Object
```js
import { cameraProjectionMatrix, highPrecisionModelViewMatrix, positionLocal } from 'three/tsl';

material = new THREE.NodeMaterial();
material.vertexNode = cameraProjectionMatrix.mul( highPrecisionModelViewMatrix).mul( positionLocal );
```

`highPrecisionModelViewMatrix` will use CPU and `modelViewMatrix` will use GPU.
`modelViewMatrix` will be the default.

I'll do the same for the normal ones.