@WestLangley 2024-09-03
__Master branch (WebGLRenderer and WebGPURenderer)__

![master](https://github.com/user-attachments/assets/62d88bc2-7db2-4481-b4d5-7c1ca489ea3d)

__This PR (WebGPURenderer)__

![169dev](https://github.com/user-attachments/assets/985dedb5-57e7-4c1b-bbf3-b9e89b426eeb)

This is because this PR computes the incorrect normals in the GPU... not surprising, based on my comments above.