@WestLangley 2024-09-03
WebGPU dev branch fiddle: https://jsfiddle.net/La1e5gmz/